requirements.txt
discord.py
python-dotenv
pytz
aiohttp
requests
psycopg2-binary
psycopg2-pool
python-dotenv